requirements.txt
python-frontmatter
ipython
